Added tests for adding appreticeships
