+++ src/SFA.DAS.Forecasting.Web.Automation/SFA.DAS.Forecasting.Web.Automation.csproj
@@ -79,7 +79,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AddApprenticeshipsToEstimateCostPage.cs" />
+    <Compile Include="Estimation\AddApprenticeshipsToEstimateCostPage.cs" />
     <Compile Include="Estimation\EditApprenticeshipsPage.cs" />
     <Compile Include="RemoveApprenticeshipPage.cs" />
     <Compile Include="EstimateCostsPage.cs" />
