+++ src/SFA.DAS.Forecasting.sln
@@ -158,7 +158,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{B6B0820B-097F-4938-A1D4-FE92F3A1308F} = {1DD9C0CC-B5FA-4244-9E57-43B8A9DCAD85}
 		{2477A59A-E3A7-4003-8C40-E48FD4E219EF} = {1DD9C0CC-B5FA-4244-9E57-43B8A9DCAD85}
 		{5EA585E1-4DA6-40BE-8611-B8F2535F7085} = {1DD9C0CC-B5FA-4244-9E57-43B8A9DCAD85}
 		{53EAF7C5-7389-4F4D-BF13-073253D32CCD} = {1DD9C0CC-B5FA-4244-9E57-43B8A9DCAD85}
